// Frontend/src/api/model/ascentResponseDTOStyle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AscentResponseDTOStyle = typeof AscentResponseDTOStyle[keyof typeof AscentResponseDTOStyle];


export const AscentResponseDTOStyle = {
  onsight: 'onsight',
  flash: 'flash',
  redpoint: 'redpoint',
  pinkpoint: 'pinkpoint',
  toprope: 'toprope',
  hangdog: 'hangdog',
  attempt: 'attempt',
} as const;